clean up package structure
