--- tonic/TODO
@@ -1,28 +1,0 @@
-
-
-** PARAMS **
-flatten input type from wosk :
-Order:
--pass path params first
--query params second (alphasort)
--body last (alphasort)
-Param names: 
--take from json/wosk.
-Resiliency to changes:
--if body params names change => api broken
--if path param change name but type stays the same, api is unchanged => to protect SDK we pass them first respection uri order
--if someone adds an optional param, sdk will be broken
-
-
-** FUNCTION NAMES **
-we don't have restful apis, so it's hard to make up meaningful names
--> we currently use handler name but:
-	-it's subject to change
-	-a merger service with different packages containing handler may provoke collisions
--> we could use some comments but's it's not super solid either
-
-but.... we could use this during version releasing and load previous version 
-  => sdk generator could ask confirmation "sdk method for POST /foobar name will change from CreateBaz to NewBaz"
-
-
-
