+++ tonic/genswagger/swagger/apideclaration.go
