+++ tonic/genswagger/swagger/model.go
