+++ tonic/genswagger/swagger/operation.go
