+++ tonic/genswagger/swagger/parameter.go
