+++ tonic/genswagger/swagger/type.go
